requirement.txt
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
Pillow==8.4.0
pytesseract==0.3.8
python-dotenv==0.19.2
scipy==1.7.3
six==1.16.0
Werkzeug==2.0.2
google-cloud-vision